refactor: extract ServiceEvent base to remove event duplication
ServiceCompleted and ServiceFailed were near-identical; extract the shared payload into an abstract ServiceEvent base they extend. Removes the duplication smell and consolidates the event generics annotation.
